Implementation of sumSeries method
Implementation of sum/sumSeries. We are creating new SeekableView for every
metric that the initial expression query contained. This might not be the
best design in terms of memory usage, but the best to guarantee that we can
re-use AggregationIterator, which does interpolation and invokes the correct
aggregation logic at the right time.

Signed-off-by: Arjun Satish <arjun.satish@turn.com>